Moved PEMEncoder Logic from SecuirtyUtils to the DDDPemEncoder class, removed an unused SecuirtyUtils.java file
