+++ bundle-core/src/main/java/net/discdd/client/bundlesecurity/ClientSecurity.java
@@ -1,5 +1,6 @@
 package net.discdd.client.bundlesecurity;
 
+import net.discdd.bundlesecurity.DDDPEMEncoder;
 import net.discdd.bundlesecurity.SecurityUtils;
 import net.discdd.pathutils.ClientPaths;
 import org.whispersystems.libsignal.DuplicateMessageException;
@@ -111,16 +112,16 @@ public class ClientSecurity {
                         ourIdentityKeyPair.getPrivateKey().serialize());
             Files.write(path.resolve(SecurityUtils.CLIENT_BASE_PRIVATE_KEY), ourBaseKey.getPrivateKey().serialize());
         }
-        SecurityUtils.createEncodedPublicKeyFile(ourIdentityKeyPair.getPublicKey().getPublicKey(), identityKeyPaths[0]);
-        SecurityUtils.createEncodedPublicKeyFile(ourBaseKey.getPublicKey(), identityKeyPaths[1]);
-        SecurityUtils.createEncodedPublicKeyFile(theirIdentityKey.getPublicKey(), identityKeyPaths[2]);
+        DDDPEMEncoder.createEncodedPublicKeyFile(ourIdentityKeyPair.getPublicKey().getPublicKey(), identityKeyPaths[0]);
+        DDDPEMEncoder.createEncodedPublicKeyFile(ourBaseKey.getPublicKey(), identityKeyPaths[1]);
+        DDDPEMEncoder.createEncodedPublicKeyFile(theirIdentityKey.getPublicKey(), identityKeyPaths[2]);
         return identityKeyPaths;
     }
 
     private void loadKeysfromFiles(Path clientKeyPath) throws IOException, InvalidKeyException {
         byte[] identityKeyPvt = Files.readAllBytes(clientKeyPath.resolve(SecurityUtils.CLIENT_IDENTITY_PRIVATE_KEY));
         byte[] identityKeyPub =
-                SecurityUtils.decodePublicKeyfromFile(clientKeyPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY));
+                DDDPEMEncoder.decodePublicKeyfromFile(clientKeyPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY));
 
         IdentityKey identityPublicKey = new IdentityKey(identityKeyPub, 0);
         ECPrivateKey identityPrivateKey = Curve.decodePrivatePoint(identityKeyPvt);
@@ -128,7 +129,7 @@ public class ClientSecurity {
         ourIdentityKeyPair = new IdentityKeyPair(identityPublicKey, identityPrivateKey);
 
         byte[] baseKeyPvt = Files.readAllBytes(clientKeyPath.resolve(SecurityUtils.CLIENT_BASE_PRIVATE_KEY));
-        byte[] baseKeyPub = SecurityUtils.decodePublicKeyfromFile(clientKeyPath.resolve(SecurityUtils.CLIENT_BASE_KEY));
+        byte[] baseKeyPub = DDDPEMEncoder.decodePublicKeyfromFile(clientKeyPath.resolve(SecurityUtils.CLIENT_BASE_KEY));
 
         ECPublicKey basePublicKey = Curve.decodePoint(baseKeyPub, 0);
         ECPrivateKey basePrivateKey = Curve.decodePrivatePoint(baseKeyPvt);
@@ -138,14 +139,14 @@ public class ClientSecurity {
 
     private void InitializeServerKeysFromFiles(Path path) throws InvalidKeyException, IOException {
         byte[] serverIdentityKey =
-                SecurityUtils.decodePublicKeyfromFile(path.resolve(SecurityUtils.SERVER_IDENTITY_KEY));
+                DDDPEMEncoder.decodePublicKeyfromFile(path.resolve(SecurityUtils.SERVER_IDENTITY_KEY));
         theirIdentityKey = new IdentityKey(serverIdentityKey, 0);
 
         byte[] serverSignedPreKey =
-                SecurityUtils.decodePublicKeyfromFile(path.resolve(SecurityUtils.SERVER_SIGNED_PRE_KEY));
+                DDDPEMEncoder.decodePublicKeyfromFile(path.resolve(SecurityUtils.SERVER_SIGNED_PRE_KEY));
         theirSignedPreKey = Curve.decodePoint(serverSignedPreKey, 0);
 
-        byte[] serverRatchetKey = SecurityUtils.decodePublicKeyfromFile(path.resolve(SecurityUtils.SERVER_RATCHET_KEY));
+        byte[] serverRatchetKey = DDDPEMEncoder.decodePublicKeyfromFile(path.resolve(SecurityUtils.SERVER_RATCHET_KEY));
         theirRatchetKey = Curve.decodePoint(serverRatchetKey, 0);
     }
